pyproject.toml
[tool.poetry]
name = 'DbObjectCreator'
version = "1.1.3"
description = 'A database object creator'
authors = ['Gary Haag <haagimus@gmail.com>']
maintainers = ['Gary Haag <haagimus@gmail.com>']
license = 'MIT'
readme = 'README.md'
homepage = ''
repository = 'https://github.com/Haagimus/DbObjectCreator'
documentation = 'https://hackersandslackers.com/python-poetry/'
keywords = ['Database',
            'MySQL',
            'MSSQL',
            'PostgreSQL',
            'Packaging',
            'ETL',
            'Data',
            'Warehouse',
            'Data Warehouse',
            ]

[tool.poetry.dependencies]
python = '^3.7'
bcrypt = '*'
cffi = '*'
cryptography = '*'
importlib_metadata = '*'
numpy = '*'
pandas = '*'
paramiko = '*'
psycopg2-binary = '*'
pycparser = '*'
pymssql = '*'
pymysql = '*'
PyNaCl = '*'
python-dateutil = '*'
pytz = '*'
six = '*'
SQLAlchemy = '*'
sshtunnel = '*'
wheel = '*'

[tool.poetry.dev-dependencies]
Cython = '*'
black = {version = "^22.3.0", allow-prereleases = true}
toml = '*'
ipykernel = "^6.15.0"

[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'

[tool.poetry.scripts]

[tool.poetry.urls]

[tool.poetry_bumpversion.file."DbObjectCreator/__init__.py"]